refactor(datagrid-web): apply review changes
